L02Q01: use tipo de dado correto para getline()
O manual (`$ man 3 getline`) exige que esse tipo numérico seja `size_t`,
e o ChatGPT me diz que isso não é o mesmo que `unsigned long` em todos
os sistemas.

+++ LAB02Q01/combinador.c
@@ -10,7 +10,7 @@ int main(void)
 {
 	char *res = NULL; // String dinamicamente alocada de resultado.
 	char *texto = NULL; // String dinamicamente alocada de entrada e saída.
-	long unsigned int cap = 0; // Capacidade da string de entrada.
+	size_t cap = 0; // Capacidade da string de entrada.
 
 	// Aloca memória e lê uma linha da entrada.
 	while (getline(&texto, &cap, stdin) != -1) {
